requirements.txt
pip
wheel
protobuf
psycopg2-binary
pydrive
pandas
boto3
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib